Check list for the next submittion
Here is what we decided on what we need to do for our web application. Style and colouring is decided by yourself and the web page only needs the things stated on the check list. the help& about us page are made after the styleing of the homepage is done.